// src/types/WooCommerceTypes.ts
export interface TAttributeTerm {
  id: number;
  name: string;
  slug: string;
  description: string;
  menu_order: string;
  count: number;
}

export interface IFilteringArray {
  id: number;
  name: string;
  slug: string;
  terms: TAttributeTerm[];
}